+++ .vscode/settings.json
@@ -71,7 +71,8 @@
         "variant": "cpp",
         "bit": "cpp",
         "dense": "cpp",
-        "forward_list": "cpp"
+        "forward_list": "cpp",
+        "valarray": "cpp"
     },
     "python.analysis.extraPaths": [
         "/workspaces/ENPM_662_Project2/devel/lib/python3/dist-packages",
